+++ src/cli.rs
@@ -1,4 +1,4 @@
-use std::{convert::TryInto, process, str::FromStr};
+use std::{convert::TryInto, env, process, str::FromStr};
 
 use clap::{crate_authors, crate_version, App, AppSettings, Arg, SubCommand};
 use color_eyre::{eyre::eyre, Help, Report};
@@ -10,7 +10,7 @@ use crate::{
 	profile::Profile,
 };
 
-pub fn try_cli(manager: &mut KeyboardManager) -> Result<bool, Report> {
+pub fn try_cli() -> Result<(), Report> {
 	let matches = App::new("Legion Keyboard Control")
 		.setting(AppSettings::ColoredHelp)
 		.version(crate_version!())
@@ -92,6 +92,8 @@ pub fn try_cli(manager: &mut KeyboardManager) -> Result<bool, Report> {
 		.get_matches();
 
 	if let Some(input) = matches.subcommand_name() {
+		let mut manager = KeyboardManager::new().unwrap();
+
 		fn parse_bytes_arg(arg: &str) -> Result<Vec<u8>, <u8 as FromStr>::Err> {
 			arg.split(',').map(str::parse::<u8>).collect()
 		}
@@ -115,7 +117,7 @@ pub fn try_cli(manager: &mut KeyboardManager) -> Result<bool, Report> {
 				if let Some(path_string) = input_matches.value_of("path") {
 					match CustomEffect::from_file(path_string.to_string()) {
 						Ok(effect) => {
-							effect.play(manager);
+							effect.play(&mut manager);
 						}
 						Err(err) => {
 							return Err(eyre!("{} ", err.to_string()).suggestion("Make sure you are using a valid effect"));
@@ -170,8 +172,11 @@ pub fn try_cli(manager: &mut KeyboardManager) -> Result<bool, Report> {
 			}
 		}
 
-		Ok(true)
+		Ok(())
 	} else {
-		Ok(false)
+		let exec_name = env::current_exe().unwrap().file_name().unwrap().to_string_lossy().into_owned();
+		println!("No subcommands found, starting in GUI mode. To view the possible subcommands type \"{} --help\".", exec_name);
+		crate::gui::app::App::start_ui();
+		Ok(())
 	}
 }
